[#3600] Fixed eigenshappen request, added tests
